# [Enhancement] optimize the performance of multi_count_distinct high cardinality string